requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy[asycnio]
aiosqlite
eyed3
Pillow
python-multipart
python-dotenv